Sound.PortMidi

Description

Interface to PortMidi

Data Types

data PMError

Represents real errors of the C enum PmError

Constructors

HostError

InvalidDeviceId

InsufficientMemory

BufferTooSmall

BufferOverflow

BadPtr

BadData

InternalError

BufferMaxSize

data PMSuccess

Represents non-errors of the C enum PmError

Constructors

NoError'NoData
Returned by poll when there is no data, and returned by other functions when there is no error.

GotData
Only returned by poll when data is available.

newtype PMEventCount

Represents a count of PMEvents

Constructors

PMEventCount CInt

type PMStream = ForeignPtr PortMidiStream

data DeviceInfo

Constructors

DeviceInfo

Fields

interface :: String

name :: String

input :: Bool

output :: Bool

opened :: Bool

type DeviceID = Int

data PMMsg

Constructors

PMMsg

Fields

status :: !CLong

data1 :: !CLong

data2 :: !CLong

data PMEvent

Constructors

PMEvent

Fields

message :: !CLong

timestamp :: !Timestamp

Constants

filterActive :: CLong

filterSysex :: CLong

filterClock :: CLong

filterPlay :: CLong

filterTick :: CLong

filterFD :: CLong

filterUndefined :: CLong

filterReset :: CLong

filterRealtime :: CLong

filterNote :: CLong

filterChannelAftertouch :: CLong

filterPolyAftertouch :: CLong

filterAftertouch :: CLong

filterProgram :: CLong

filterControl :: CLong

filterPitchBend :: CLong

filterMTC :: CLong

filterSongPosition :: CLong

filterSongSelect :: CLong

filterTune :: CLong

filterSystemCommon :: CLong

PortMidi functions

initialize :: IO (Either PMError PMSuccess)

terminate :: IO (Either PMError PMSuccess)

hasHostError :: PMStream -> IO Bool

getErrorText :: PMError -> IO String

getSuccessText :: PMSuccess -> IO String

getText :: Either PMError PMSuccess -> IO String

countDevices :: IO DeviceID

getDefaultInputDeviceID :: IO (Maybe DeviceID)

getDefaultOutputDeviceID :: IO (Maybe DeviceID)

getDeviceInfo :: DeviceID -> IO DeviceInfo

openInput :: DeviceID -> IO (Either PMError PMStream)

openOutput :: DeviceID -> Int -> IO (Either PMError PMStream)

setFilter :: PMStream -> CLong -> IO (Either PMError PMSuccess)

channel :: Int -> CLong

setChannelMask :: PMStream -> CLong -> IO (Either PMError PMSuccess)

abort :: PMStream -> IO (Either PMError PMSuccess)

close :: PMStream -> IO (Either PMError PMSuccess)

poll :: PMStream -> IO (Either PMError PMSuccess)

Returns wether or not a subsequent call to readEvents would return some PMEvents or not.

readEvents :: PMStream -> IO (Either PMError [PMEvent])

Reads at most 256 PMEvents, using a dynamically allocated buffer.

readEventsToBuffer

Arguments

:: PMStream

-> Ptr PMEvent
The PMEvents buffer which will contain the results.

-> CLong
The size of the PMEvent buffer, in number of elements. No more that this number of PMEvents can be read at once.

-> IO (Either PMError PMEventCount)
When Right, returns the number of elements written to the PMEvent buffer.

Reads PMEvents and writes them to the buffer passed as argument.

writeEvents :: PMStream -> [PMEvent] -> IO (Either PMError PMSuccess)

writeShort :: PMStream -> PMEvent -> IO (Either PMError PMSuccess)

writeSysEx :: PMStream -> Timestamp -> String -> IO (Either PMError PMSuccess)

Time function

time :: IO Timestamp

encodeMsg :: PMMsg -> CLong

decodeMsg :: CLong -> PMMsg